// types/plex-metadata.ts
export interface PlexMetadata {
  MediaContainer: {
    Metadata: {
      guid: string; // com.plexapp.agents.thetvdb://260449/4/18?lang=en
      type: string; // episode
      title: string; // Revenge
      index: number; // 18
      parentIndex: number; // 4
      viewOffset: number; // 104519,
      year: number; // 2017
      duration: number; // 2610149;
    }[];
  };
};